moved game Scratch Card to the games package
