+++ gazeplay/src/main/java/net/gazeplay/games/scratchcard/ScratchcardGamesStats.java
@@ -1,6 +1,7 @@
-package utils.games.stats;
+package net.gazeplay.games.scratchcard;
 
 import javafx.scene.Scene;
+import utils.games.stats.HiddenItemsGamesStats;
 
 public class ScratchcardGamesStats extends HiddenItemsGamesStats {
